transfer attraction endpoint to attractionController
