+++ routes/userRoute.js
@@ -6,11 +6,13 @@ const {
   login,
   register,
   dashboard,
-  token} = require('../controllers/userController');
+  token,
+  logout} = require('../controllers/userController');
 
 router.post('/login', login);
 router.post('/register', register);
 router.post('/token', token);
 router.post('/dashboard', authenticateToken, dashboard);
+router.post('/logout', logout);
 
 module.exports = router;
